Blockade generation and collision checking
This commit sees automatic blockade generation and improved collision checking. Blockades are now automatically, and randomly, assigned to the Board, with some parameters in place. Collision checking previously did not effectively catch when the robot would hit the grid border, and was not reliable for blockades, as well. These issues have been eliminated, and any hit blockades now turn a true red color.